LPS-76269 Delete html preview only if it exist

+++ modules/apps/web-experience/fragment/fragment-service/src/main/java/com/liferay/fragment/service/impl/FragmentEntryLocalServiceImpl.java
@@ -141,8 +141,10 @@ public class FragmentEntryLocalServiceImpl
 
 		// HTML preview
 
-		_htmlPreviewEntryLocalService.deleteHtmlPreviewEntry(
-			fragmentEntry.getHtmlPreviewEntryId());
+		if (fragmentEntry.getHtmlPreviewEntryId() > 0) {
+			_htmlPreviewEntryLocalService.deleteHtmlPreviewEntry(
+				fragmentEntry.getHtmlPreviewEntryId());
+		}
 
 		return fragmentEntry;
 	}
